補update程式碼露加

+++ delete-article.php
@@ -27,18 +27,20 @@
         die("id 不存在, 文章未發現");
     }
 
-    $sql = "DELETE FROM article 
-            WHERE id = ?";
+    if ($_SERVER["REQUEST_METHOD"] === "POST") {
+        $sql = "DELETE FROM article 
+                WHERE id = ?";
 
-    $stmt = mysqli_prepare($conn, $sql);
-    if($stmt === false){
-        echo mysqli_error($conn);
-        exit;
-    }
-    mysqli_stmt_bind_param($stmt, "i", $id);
+        $stmt = mysqli_prepare($conn, $sql);
+        if($stmt === false){
+            echo mysqli_error($conn);
+            exit;
+        }
+        mysqli_stmt_bind_param($stmt, "i", $id);
 
-    if(mysqli_stmt_execute($stmt)){
-        redirect("/PHP-beginger/index.php");
-    }else{
-        echo mysqli_stmt_error($stmt);
+        if(mysqli_stmt_execute($stmt)){
+            redirect("/PHP-beginger/index.php");
+        }else{
+            echo mysqli_stmt_error($stmt);
+        }
     }
